+++ cmd/influxd/upgrade/upgrade.go
@@ -29,6 +29,7 @@ import (
 	"github.com/influxdata/influxdb/v2/v1/services/meta/filestore"
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
 )
 
 // Simplified 1.x config.
@@ -112,7 +113,9 @@ var options = struct {
 	// verbose output
 	verbose bool
 
-	logPath string
+	// logging
+	logLevel string
+	logPath  string
 
 	force bool
 }{}
@@ -234,6 +237,12 @@ func NewCommand() *cobra.Command {
 			Desc:    "optional: Custom InfluxDB 1.x config file path, else the default config file",
 		},
 		{
+			DestP:   &options.logLevel,
+			Flag:    "log-level",
+			Default: zapcore.InfoLevel.String(),
+			Desc:    "supported log levels are debug, info, warn and error",
+		},
+		{
 			DestP:   &options.logPath,
 			Flag:    "log-path",
 			Default: filepath.Join(homeOrAnyDir(), "upgrade.log"),
@@ -299,8 +308,14 @@ func runUpgradeE(*cobra.Command, []string) error {
 		fluxInitialized = true
 	}
 
+	var lvl zapcore.Level
+	if err := lvl.Set(options.logLevel); err != nil {
+		return errors.New("unknown log level; supported levels are debug, info, warn and error")
+	}
+
 	ctx := context.Background()
 	config := zap.NewProductionConfig()
+	config.Level = zap.NewAtomicLevelAt(lvl)
 	config.OutputPaths = append(config.OutputPaths, options.logPath)
 	config.ErrorOutputPaths = append(config.ErrorOutputPaths, options.logPath)
 	log, err := config.Build()