+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 1. [19960](https://github.com/influxdata/influxdb/pull/19960): Remove bucket and mapping auto-creation from v1 /write API
 1. [19972](https://github.com/influxdata/influxdb/pull/19972): Remove unused 'influx-command-path' option from upgrade command
 1. [19969](https://github.com/influxdata/influxdb/pull/19969): Check if CLI configs file already exists during upgrade
+1. [19967](https://github.com/influxdata/influxdb/pull/19967): Add 'log-level' option to upgrade command
 
 ## v2.0.0-rc.4 [2020-11-05]